[MLPerf GPT3] Add GPT3 model configuration

+++ MaxText/max_utils.py
@@ -240,8 +240,8 @@ def init_initial_state(model, tx, config, is_training, key):
       config.max_target_length
   )
   model_vars = model.init({'params': key, 'dropout': key, 'aqt': key},
-                          jnp.ones(input_shape),
-                          jnp.ones(input_shape))
+                          jnp.ones(input_shape, dtype=jnp.int32),
+                          jnp.ones(input_shape, dtype=jnp.int32))
   if is_training:
     return init_training_state(model.apply, model_vars['params'], tx)
   return init_decode_state(model.apply, model_vars['params'])
